Added FX changes and int->double for amount

+++ src/User.cpp
@@ -13,7 +13,7 @@ User::User(string name){
 		this->total_owes=0;
 	}
 
-User::User(string name, int balance, int total_owed, int total_owes, unordered_map<string,int> tagCount){
+User::User(string name, double balance, double total_owed, double total_owes, unordered_map<string,int> tagCount){
 	this->name=name;
 	this->balance=balance;
 	this->total_owed=total_owed;
@@ -27,15 +27,15 @@ void User::getBalanceDetails(){
 	cout<<"Net balance for "<<name<<" is "<<balance<<"\n";
 }
 
-int User::getTotalOwes(){
+double User::getTotalOwes(){
 	return this->total_owes;
 }
 
-int User::getTotalOwed(){
+double User::getTotalOwed(){
 	return this->total_owed;
 }
 
-int User::getBalance(){
+double User::getBalance(){
 	return this->balance;
 }
 
@@ -54,8 +54,8 @@ int User::getTotalUnits(vector<string> ratios){
 	return sum;
 }
 //Expense when some other spends and this user is involved
-void User::addOwesExpense(string name, int multiplier, int perUnitAmount, string tag){
-	int owesToUser = multiplier*perUnitAmount;
+void User::addOwesExpense(string name, int multiplier, double perUnitAmount, string tag){
+	double owesToUser = multiplier*perUnitAmount;
 	transactions[name].push_back(to_string(-1*(owesToUser)));
 	total_owes+=owesToUser;
 	balanceOut();
@@ -65,13 +65,13 @@ void User::addOwesExpense(string name, int multiplier, int perUnitAmount, string
 		tagCount[tag]=owesToUser;
 }
 //Expense when this user spends
-void User::addOwedExpense(vector<string> usersToBeSplitWith,vector<string> ratios,int amount,string tag){
+void User::addOwedExpense(vector<string> usersToBeSplitWith,vector<string> ratios,double amount,string tag){
 	int totalUnits = getTotalUnits(ratios);
 	cout<<"Total units are: "<<totalUnits<<"\n";
-	int perUnitAmount=amount/totalUnits;
+	double perUnitAmount=amount/totalUnits;
 	cout<<"Amount is "<<amount<<" and per unit amount is "<<perUnitAmount<<"\n";
 	for(unsigned int i=1;i<ratios.size();i++){
-		int owedByThis = perUnitAmount*stoi(ratios[i]);
+		double owedByThis = perUnitAmount*stoi(ratios[i]);
 		total_owed+=owedByThis;
 		string owedByThisUser= to_string(owedByThis);
 		try{
@@ -90,7 +90,7 @@ void User::addOwedExpense(vector<string> usersToBeSplitWith,vector<string> ratio
 }
 
 int User::simplifyTransactions(string name){
-	int finalAmount=0;
+	double finalAmount=0;
 	vector<string> transactionsWithThisUser = transactions[name];
 	for(auto x:transactionsWithThisUser)
 		finalAmount+=stoi(x);
@@ -100,7 +100,7 @@ void User::simplify(){
 	cout<<"\nStarting the netting of transactions for "<<name<<":\n";
 	for(auto x:transactions){
 		cout<<"Transactions with "<<x.first<<"\n";
-		int answer = simplifyTransactions(x.first);
+		double answer = simplifyTransactions(x.first);
 		cout<<"For user "<<x.first<<", the net amount owed/owes is "<<answer<<"\n";
 	}
 }
@@ -110,12 +110,12 @@ void User::expenditureByTag(string tag){
 	else
 		cout<<"The user does not have transaction involved with this tag!\n";
 }
-void User::makePaymentToUser(string name, int amount){
+void User::makePaymentToUser(string name, double amount){
 	transactions[name].push_back(to_string(amount));
 	total_owes-=amount;
 	balanceOut();
 }
-void User::receivePaymentFromUser(string name, int amount){
+void User::receivePaymentFromUser(string name, double amount){
 	transactions[name].push_back(to_string(-1*amount));
 	total_owed-=amount;
 	balanceOut();
